+++ src/extensions/super/superext_hooks.cpp
@@ -31,9 +31,62 @@
 #include "fatal.h"
 #include "asserthandler.h"
 #include "debughandler.h"
+#include "extension.h"
+#include "unit.h"
+#include "house.h"
+#include "housetype.h"
+#include "sideext.h"
 
 #include "hooker.h"
 #include "hooker_macros.h"
+
+
+/**
+ *  Helper function that creates a hunter-seeker for the house's side.
+ *
+ *  @author: ZivDero
+ */
+static UnitClass* Make_HunterSeeker(HouseClass* house)
+{
+    const auto side_ext = Extension::Fetch<SideClassExtension>(Sides[house->Class->Side]);
+
+    if (side_ext->HunterSeeker) {
+        return new UnitClass(const_cast<UnitTypeClass*>(side_ext->HunterSeeker), house);
+    }
+
+    return nullptr;
+}
+
+
+/**
+ *  Patch to use the hunter-seeker for the house's side.
+ *
+ *  @author: ZivDero
+ */
+DECLARE_PATCH(_SuperClass_Place_HunterSeeker_Type_Patch)
+{
+    GET_REGISTER_STATIC(SuperClass*, this_ptr, esi);
+    static UnitClass* hunter_seeker;
+
+    /**
+     *  Fetch the hunter-seeker for this house's side.
+     */
+    hunter_seeker = Make_HunterSeeker(this_ptr->House);
+    _asm mov esi, hunter_seeker
+
+    /**
+     *  If we've successfully created a hunter-seeker, proceed to launching it.
+     */
+    if (hunter_seeker) {
+        JMP(0x0060C642);
+    }
+    /**
+     *  Otherwise, abort (return).
+     */
+    else {
+        JMP(0x0060C68F);
+    }
+}
 
 
 /**
@@ -45,4 +98,6 @@ void SuperClassExtension_Hooks()
      *  Initialises the extended class.
      */
     SuperClassExtension_Init();
+
+    Patch_Jump(0x0060C5DE, &_SuperClass_Place_HunterSeeker_Type_Patch);
 }